move reload logic out
